DEV: Update controller:topic modifications with native class syntax
In preparation for https://github.com/discourse/discourse/pull/28224
